requirements.txt
aiofiles==23.2.1
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip#sha256=48eeeb58b7c331ac6dec0f3533c1fb0f67f7119486bfa0482796d71ec73306c3
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
packaging==23.2
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
redis==5.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
yarl==1.9.2
